Tweaked genetic algorithm to be good again
I was hoping that the polynomial weights would automatically fix the
issue where two of the stattypes have much smaller possible ranges for
their value increase thus making them less attractive in a simple
constant * value calculation. However it seems that the algorithm
still has issues with finding good solutions which use these values.

Maybe that is to be expected since there it is difficult for a gene
that goes in the direction of using crit dmg for example. The distance
needed to travel is very long and does not give any benefits until it
has put a lot of weight on that stat type.

Also changed the mutation rate and mutation amplitude. The previous
values caused the algorithm to quite go away from already good weights
to bad ones. Not sure why it would not find its way back to good
values. Will have to investigate more.