Added skill tree to player
Have the old and new way of representing skills in Hero now, update
the old stuff in the ChooseRandomSkill method later.

A playable hero is created which has a skillTreeGenerator which
generates skills which its AI think are evenly matched.

Adjusted the player UI a bit to cater for this. The values for skills
are now doubles and you dont get to choose every option every time
which put some new requirements on the UI.

 * Need to see your stats all the time

 * Dont show the whole stat values, truncate/round to 2 decimal points

Not super interesting yet since there is no randomness in the options.