Added skilltree to evolution
So now it is possible to run evolution when new skills are generated
from a generator instead of being random. This gives the option to do
a second round of evolution and see how well the second gen does when
the stats are balanced by the first.

Also put the question of adding a human player later in the UI so its
asked when its needed.