Create IGame interface
The goal is to put basically all what is now in Program.cs into game
classes. Then when building for release we have a PlayerGame and when
in debug we can have an AIGame, PlayerGame or whatever.

This will make Program.cs much cleaner, and it makes it easier to
debug and demo things.