All classes use IConsole now
These changes were also neccesary, but dont feel as good as the previous
commit. Mostly since PlayerGame and AIGame are not needed anymore it
feels. They could both be PlayerGame but with different IConsoles.

Also there are places where IConsole seems crammed in right now. This is
probably since there are console calls everywhere. A cleanup is needed.

Review this commit later on when looking at where to change how IConsole
is used. Right now the focus was simply to get everythin back to working
as before.