Added Singelton to get IConsole from
This change greatly improved the architecture of the program. There is
much less sending extra arguments around and less properties in all
objects to understand.

It is still possible and easy to change the console to be used in
tests.

The only downside is that it is no longer possible to have as much
low-lever control over if a particular method is allowed to use what
console. If you want to change what console a particular method uses
you have to change the global console which is sub-optimal. But it is
unclear if that really is something that will be an issue.