Updated Dockerfile to include test step
When building a dockerfile it is also tested now. Hopefully this can
be integrated with the automatic build tools to check for errors.

Also the hope is that the execution times for the tests will be logged
so that performance issues can be tracked down to individual commits
in the future.