Added Random to the DataHub
All random accesses go through the DataHub. This was done mainly from
a code-maintainability standpoint. Not having to send random as a
parameter to most new objects is good for readability.

However it should also make it easier to get determenistic results
since only one instance of Random is used and the seed of it can be
controled.

There was a hope that it would result in a speed-up as well but no
speed-up was noted. The previous solution did only create a handful
random objects or so so thats probably the reason.